Template fields: 'number' type now accepts decimals (was integer-only)
Three places treated 'number' field type as INT, causing 2.2 to be rejected:

1. EntryForm validation: case 'number' added 'integer' rule. Now falls
   through to 'numeric' (same as 'decimal'/'float'). Matches HTML5
   input[type=number] behaviour where decimals are allowed by default.

2. TemplateTableGenerator column type: case 'number' used ->integer().
   Now uses ->decimal($name, 12, 2) — same as 'decimal'/'float'. Range
   covers up to 9,999,999,999.99 which is more than enough for areas /
   prices.

3. TemplateTableGenerator model cast: case 'number' cast to 'integer'.
   Now casts to 'float' so reading 2.20 back yields 2.2.

Migration alter_project_tables_size_columns_to_decimal converts the
existing INT columns (building_size, plot_size, pool_size) to
DECIMAL(12,2) for completed_villas, under_constructions, renovations.
year_built stays INT — a year is integer.

Updated CompletedVilla / UnderConstruction / Renovation model casts.

If you need INT-only validation, use field type 'integer' explicitly.

+++ app/Models/UnderConstruction.php
@@ -20,9 +20,9 @@ class UnderConstruction extends Model implements HasMedia
     protected $casts = [
         'gallery' => 'array',
         'year_built' => 'integer',
-        'building_size' => 'integer',
-        'plot_size' => 'integer',
-        'pool_size' => 'integer',
+        'building_size' => 'float',
+        'plot_size' => 'float',
+        'pool_size' => 'float',
     ];
     /**
      * Get the URL for Main Photo
